--- a/src/tensors/combinators/permissive.jl
+++ b/src/tensors/combinators/permissive.jl
@@ -119,6 +119,7 @@ end
 jumper_body(node::VirtualPermissiveArray, ctx, ext) = VirtualPermissiveArray(jumper_body(node.body, ctx, ext), node.dims)
 stepper_body(node::VirtualPermissiveArray, ctx, ext) = VirtualPermissiveArray(stepper_body(node.body, ctx, ext), node.dims)
 stepper_seek(node::VirtualPermissiveArray, ctx, ext) = stepper_seek(node.body, ctx, ext)
+jumper_seek(node::VirtualPermissiveArray, ctx, ext) = jumper_seek(node.body, ctx, ext)
 
 getroot(tns::VirtualPermissiveArray) = getroot(tns.body)