--- a/src/tensors/combinators/protocolized.jl
+++ b/src/tensors/combinators/protocolized.jl
@@ -124,9 +124,9 @@ function unfurl(tns::VirtualProtocolizedArray, ctx, ext, protos...)
     VirtualProtocolizedArray(unfurl(tns.body, ctx, ext, map(something, tns.protos, protos)...), tns.protos)
 end
 
-jumper_body(node::VirtualProtocolizedArray, ctx, ext) = VirtualProtocolizedArray(jumper_body(node.body, ctx, ext), node.protos)
-stepper_body(node::VirtualProtocolizedArray, ctx, ext) = VirtualProtocolizedArray(stepper_body(node.body, ctx, ext), node.protos)
-stepper_seek(node::VirtualProtocolizedArray, ctx, ext) = stepper_seek(node.body, ctx, ext)
-jumper_seek(node::VirtualProtocolizedArray, ctx, ext) = jumper_seek(node.body, ctx, ext)
+#stepper_body(node::VirtualProtocolizedArray, ctx, ext) = VirtualProtocolizedArray(stepper_body(node.body, ctx, ext), node.protos)
+#stepper_seek(node::VirtualProtocolizedArray, ctx, ext) = stepper_seek(node.body, ctx, ext)
+replay_body(node::VirtualProtocolizedArray, ctx, ext) = VirtualProtocolizedArray(replay_body(node.body, ctx, ext), node.protos)
+replay_seek(node::VirtualProtocolizedArray, ctx, ext) = replay_seek(node.body, ctx, ext)
 
 getroot(tns::VirtualProtocolizedArray) = getroot(tns.body)