--- a/src/tensors/combinators/toeplitz.jl
+++ b/src/tensors/combinators/toeplitz.jl
@@ -116,10 +116,10 @@ visit_simplify(node::VirtualToeplitzArray) = VirtualToeplitzArray(visit_simplify
     guard => VirtualToeplitzArray(body, node.dim)
 end
 
-jumper_body(node::VirtualToeplitzArray, ctx, ext) = VirtualToeplitzArray(jumper_body(node.body, ctx, ext), node.dim)
-stepper_body(node::VirtualToeplitzArray, ctx, ext) = VirtualToeplitzArray(stepper_body(node.body, ctx, ext), node.dim)
-stepper_seek(node::VirtualToeplitzArray, ctx, ext) = stepper_seek(node.body, ctx, ext)
-jumper_seek(node::VirtualToeplitzArray, ctx, ext) = jumper_seek(node.body, ctx, ext)
+#stepper_body(node::VirtualToeplitzArray, ctx, ext) = VirtualToeplitzArray(stepper_body(node.body, ctx, ext), node.dim)
+#stepper_seek(node::VirtualToeplitzArray, ctx, ext) = stepper_seek(node.body, ctx, ext)
+replay_body(node::VirtualToeplitzArray, ctx, ext) = VirtualToeplitzArray(replay_body(node.body, ctx, ext), node.dim)
+replay_seek(node::VirtualToeplitzArray, ctx, ext) = replay_seek(node.body, ctx, ext)
 
 getroot(tns::VirtualToeplitzArray) = getroot(tns.body)