--- a/src/tensors/combinators/unfurled.jl
+++ b/src/tensors/combinators/unfurled.jl
@@ -104,6 +104,7 @@ end
 jumper_body(node::Unfurled, ctx, ext) = Unfurled(node.arr, node.ndims, jumper_body(node.body, ctx, ext))
 stepper_body(node::Unfurled, ctx, ext) = Unfurled(node.arr, node.ndims, stepper_body(node.body, ctx, ext))
 stepper_seek(node::Unfurled, ctx, ext) = stepper_seek(node.body, ctx, ext)
+jumper_seek(node::Unfurled, ctx, ext) = jumper_seek(node.body, ctx, ext)
 
 function lower(node::Unfurled, ctx::AbstractCompiler, ::DefaultStyle)
     ctx(node.body)