--- a/src/tensors/combinators/windowed.jl
+++ b/src/tensors/combinators/windowed.jl
@@ -122,6 +122,7 @@ end
 jumper_body(node::VirtualWindowedArray, ctx, ext) = VirtualWindowedArray(jumper_body(node.body, ctx, ext), node.dims)
 stepper_body(node::VirtualWindowedArray, ctx, ext) = VirtualWindowedArray(stepper_body(node.body, ctx, ext), node.dims)
 stepper_seek(node::VirtualWindowedArray, ctx, ext) = stepper_seek(node.body, ctx, ext)
+jumper_seek(node::VirtualWindowedArray, ctx, ext) = jumper_seek(node.body, ctx, ext)
 
 getroot(tns::VirtualWindowedArray) = getroot(tns.body)