@nullplay 2023-09-06
Some lessons I've learned & future works:

1. Distinguish a prologue/main/epilogue in `replay` for better code optimization
2. Clever `Switch` looplet lowering with bounds inference will remove unnecessary branches generated from `Step` or `Jump`. Below is the example of `for i = _; s[] += x[gallop(i)]` where `x` is `SparseList`. 
```
    while i <= phase_stop
            x_lvl_i2 = x_lvl.idx[x_lvl_q]
            phase_stop_2 = min(x_lvl_i2, phase_stop)
            if phase_stop_2 >= i
                if x_lvl_i2 == phase_stop_2  <<<<<<<<<<<<<<<<< This condition is always true if there is a single operand.
                    x_lvl_2_val_2 = x_lvl_2.val[x_lvl_q]
                    s_val = x_lvl_2_val_2
                    x_lvl_q += 1
               % else     <<<<<<<<<<<<<<<<< WE DONT WANT THIS BRANCH
               %    while i <= phase_stop_2
               %         x_lvl_i2 = x_lvl.idx[x_lvl_q]
               %         phase_stop_3 = min(x_lvl_i2, phase_stop_2)
               %         if x_lvl_i2 == phase_stop_3
               %             x_lvl_2_val_2 = x_lvl_2.val[x_lvl_q]
               %             s_val = x_lvl_2_val_2
               %             x_lvl_q += 1
               %         end
               %         i = phase_stop_3 + 1
               %     end
               % end
                i = phase_stop_2 + 1
            end
        end
```
3. Would there be a clever way to emit below code without relying on `Thunk`? Below code is coming from `for i = _; s[] += x[i]` where x is `SparseVBL`. The level definition of SparseVBL does not use `next` field in `Step` to avoid the use of `Switch`. Instead, `Thunk` with `epilogue` enables generating a clean way to increment position : `y_lvl_r += phase_stop2 == y_lvl_i`. My question is : Can we emit such code with only `next` and without calling `Thunk`? 
```
        while i <= phase_stop
            y_lvl_i = y_lvl.idx[y_lvl_r]
            y_lvl_q_stop = y_lvl.ofs[y_lvl_r + 1]
            y_lvl_q_ofs = (y_lvl_q_stop - y_lvl_i) - 1
            phase_stop_2 = min(phase_stop, y_lvl_i)
            y_lvl_r += phase_stop_2 == y_lvl_i    <<<< Such a clean way to increment position 
            i = phase_stop_2 + 1
        end
```