@nullplay 2023-09-06
@willow-ahrens I anticipate that the interface for `Replay`, `Step`, and `Jump` may evolve as I continue to refine the lowering mechanism, as explained above. Could we proceed with merging this into the main branch first? We can then address documentation once we've finalized the `Replay` lowering process.

+ Are you satisfied with the name `Replay`? We no longer have separate `Stepper` or `Jumper` looplet; instead, we've introduced a unified interface called `Replay`, where the `Replay` body can be either `Step` or `Jump`.