@willow-ahrens 2023-09-06
I'm not worried so much about the name, I think Fred suggested "Repeater" in the past, which has a similar name. We can skip docs for now. My main concern is whether this change represents a performance or code clarity regression. If we feel that this change does not decrease performance or code clarity, we can merge it. Have you tried running benchmarks to check for regressions? Theres a script called `runjudge.jl` that compares a feature branch against main automatically and generates a report