@nullplay 2023-09-06
# Performance 

## This PR
```
                                                          ID            time    GC time          memory allocations
  –––––––––––––––––––––––––––––––––––––––––––––––––––––––––– ––––––––––––––– –––––––––– ––––––––––––––– –––––––––––
                               ["compile", "compile_SpGeMM"]  38.716 ms (5%)             33.38 MiB (1%)      938625
                         ["compile", "time_to_first_SpGeMM"]    4.591 s (5%)             66.72 KiB (1%)          72
              ["graphs", "bellmanford", "Newman/netscience"]   2.958 μs (5%)             54.33 KiB (1%)          29
                ["graphs", "bellmanford", "SNAP/roadNet-CA"] 143.719 ms (5%)             94.37 MiB (1%)        3265
                     ["graphs", "bfs", "SNAP/soc-Epinions1"]   3.299 ms (5%)              4.83 MiB (1%)          44
                  ["graphs", "bfs", "SNAP/soc-LiveJournal1"] 455.126 ms (5%)            190.80 MiB (1%)          61
                ["graphs", "pagerank", "SNAP/soc-Epinions1"]  20.664 ms (5%)             12.77 MiB (1%)          34
             ["graphs", "pagerank", "SNAP/soc-LiveJournal1"]    3.804 s (5%) 666.000 μs   2.15 GiB (1%)          50
                ["indices", "SpMV_32", "SNAP/soc-Epinions1"]   1.676 ms (5%)            593.00 KiB (1%)           2
                ["indices", "SpMV_64", "SNAP/soc-Epinions1"]   1.663 ms (5%)            593.00 KiB (1%)           2
  ["matrices", "ATA_spgemm_gustavson", "SNAP/soc-Epinions1"]    1.386 s (5%)            812.63 MiB (1%)       27485
      ["matrices", "ATA_spgemm_outer", "SNAP/soc-Epinions1"]    5.612 s (5%)  32.662 ms   3.78 GiB (1%)         181
```

## Main Branch
```
                                                          ID            time   GC time          memory allocations
  –––––––––––––––––––––––––––––––––––––––––––––––––––––––––– ––––––––––––––– ––––––––– ––––––––––––––– –––––––––––
                               ["compile", "compile_SpGeMM"]  38.410 ms (5%)            33.38 MiB (1%)      938625
                         ["compile", "time_to_first_SpGeMM"]    4.553 s (5%)            66.72 KiB (1%)          72
              ["graphs", "bellmanford", "Newman/netscience"]   2.953 μs (5%)            54.33 KiB (1%)          29
                ["graphs", "bellmanford", "SNAP/roadNet-CA"] 144.848 ms (5%)            94.37 MiB (1%)        3265
                     ["graphs", "bfs", "SNAP/soc-Epinions1"]   3.351 ms (5%)             4.83 MiB (1%)          44
                  ["graphs", "bfs", "SNAP/soc-LiveJournal1"] 462.471 ms (5%)           190.80 MiB (1%)          61
                ["graphs", "pagerank", "SNAP/soc-Epinions1"]  20.914 ms (5%)            12.77 MiB (1%)          34
             ["graphs", "pagerank", "SNAP/soc-LiveJournal1"]    3.869 s (5%) 23.858 ms   2.15 GiB (1%)          50
                ["indices", "SpMV_32", "SNAP/soc-Epinions1"]   1.663 ms (5%)           593.00 KiB (1%)           2
                ["indices", "SpMV_64", "SNAP/soc-Epinions1"]   1.707 ms (5%)           593.00 KiB (1%)           2
  ["matrices", "ATA_spgemm_gustavson", "SNAP/soc-Epinions1"]    1.412 s (5%) 13.598 ms 812.63 MiB (1%)       27485
      ["matrices", "ATA_spgemm_outer", "SNAP/soc-Epinions1"]    5.594 s (5%) 28.861 ms   3.78 GiB (1%)         181

```

The performance seems identical. 

# Code clarity
## This PR
```
jaeyeonwon:/Users/jaeyeonwon/Finch.jl/test/reference64 $ wc -l typical_merge_gallop.txt                                                                                                                               3:14PM
     341 typical_merge_gallop.txt
jaeyeonwon:/Users/jaeyeonwon/Finch.jl/test/reference64 $ wc -l typical_merge_leadfollow.txt                                                                                                                           3:14PM
     246 typical_merge_leadfollow.txt
```

## Main Branch
```
jaeyeonwon:/Users/jaeyeonwon/Finch.jl/test/reference64 $ wc -l typical_merge_gallop.txt                                                                                                                               3:14PM
     365 typical_merge_gallop.txt
jaeyeonwon:/Users/jaeyeonwon/Finch.jl/test/reference64 $ wc -l typical_merge_leadfollow.txt                                                                                                                           3:14PM
     258 typical_merge_leadfollow.txt
```

10-20 less lines of generated code compared to main.