@willow-ahrens 2023-09-06
I think this code is not emitting galloping intersection.